# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {
            "name": "MJ Ma",
            "url": "https://www.linkedin.com/in/mj-profile/",
            "email": "m.mjw.ma@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/shellingford37/jump-jump/blob/master/LICENSE"
        }
    },
    "paths": {
...
        "/user/change-password": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "修改账号密码",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "账号"
                ],
                "summary": "修改账号密码",
                "parameters": [
                    {
                        "description": "修改密码请求",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ChangePasswordAPIRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Response"
                        }
                    },
                    "401": {
                        "description": ""
                    }
                }
            }
        },
        "/user/info": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "获取账号信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "账号"
                ],
                "summary": "获取账号信息",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/GetUserInfoAPIResponseData"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": ""
                    }
                }
            }
        },
        "/user/login": {
            "post": {
                "description": "账号密码登入",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "账号"
                ],
                "summary": "账号登入",
                "parameters": [
                    {
                        "description": "登入请求",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/LoginAPIRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/LoginAPIResponseData"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/user/logout": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "登出",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "账号"
                ],
                "summary": "登出",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Response"
                        }
                    },
                    "401": {
                        "description": ""
                    }
                }
            }
        }
    },
    "definitions": {
        "ChangePasswordAPIRequest": {
            "type": "object",
            "properties": {
                "newPassword": {
                    "description": "新密码",
                    "type": "string"
                },
                "password": {
                    "description": "原密码",
                    "type": "string"
                }
            }
        },
...
        "GetUserInfoAPIResponseData": {
            "type": "object",
            "properties": {
                "role": {
                    "description": "角色，1 普通用户 | 2 管理员",
                    "type": "integer",
                    "enum": [
                        1,
                        2
                    ],
                    "example": 1
                },
                "username": {
                    "description": "用户名",
                    "type": "string",
                    "example": "admin"
                }
            }
        },
...
        "LoginAPIRequest": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "description": "密码",
                    "type": "string",
                    "example": "your_password"
                },
                "username": {
                    "description": "用户名",
                    "type": "string",
                    "example": "your_username"
                }
            }
        },
        "LoginAPIResponseData": {
            "type": "object",
            "properties": {
                "token": {
                    "description": "json web token",
                    "type": "string",
                    "example": "xxx.xxx.xxx"
                }
            }
        },
...
        "Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "format": "int",
                    "default": 0,
                    "example": 0
                },
                "data": {
                    "type": "object"
                },
                "msg": {
                    "type": "string",
                    "default": "ok",
                    "example": "ok"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}